requirements.txt
fastapi~=0.115.0
starlette>=0.40.0
pydantic~=2.8.2
uvicorn[standard]~=0.21.0
pandas~=2.2.3
pillow~=11.1.0
